+++ jmeter-service/go.mod
@@ -5,7 +5,7 @@ go 1.16
 require (
 	github.com/cloudevents/sdk-go/v2 v2.6.1
 	github.com/kelseyhightower/envconfig v1.4.0
-	github.com/keptn/go-utils v0.10.1-0.20211102080258-6ac1aa413ca0
+	github.com/keptn/go-utils v0.10.1-0.20211111113359-c61381f34091
 	github.com/sirupsen/logrus v1.8.1
 	gopkg.in/yaml.v3 v3.0.0-20210107192922-496545a6307b
 )
